requirements.txt
asyncio
python-dotenv
pydub
moviepy
azure-storage-file-datalake
aiofiles
aiohttp
requests